backend/requirements.txt
anthropic>=0.51.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
twilio>=8.12.0
requests>=2.31.0
# File processing
pypdf>=4.2.0
python-docx>=1.1.0
openpyxl>=3.1.2
